requirements.txt
flet
flet-cli
flet-core
flet-desktop
flet-runtime
flet-web
google-cloud-storage
google-auth
datetime